Added a detector rosnode in line_detection.
A node was added that offers a service to detect lines in an image.
The reason I implemented it as a service and not via messages is
that it makes it easy this way, to make shure the requesting node
knows always to which image the line message corresponds.
Additionally, a node was created to test the detector node by sending
and displaying the images.