Most up to date. All changes are included here.
I made a little mess with what was changed where. I did the changes of
the last pull request in the branch 3d_line_toolbox_v0 and tried to get
rid of commits that were independent of the pull request I opened. But I
failed to do it properly, so all changes are now within this commit.