Fix the problem of mistaking surface lines as discontinuty lines.
Add "shrink2Dlines" function to lighten the influence of the bad
start or end points detected in 2D image.

Add tracking of the 2D lines associated with 3D lines and one can use
"printToFile" function to print them to txt files.

Remove the 2D lines that have small amount of points in the rectangle patches.