Remove newly-added function for convexity determination and add new one
The previously added function was not doing the right type of
geometrical considerations: it is not possible, without looking at the
inlier points, to determine whether the two planes form a convex or
concave angle when seen from a viewpoint. Indeed the two planes, by
intersecting at the line, each divide the other plane in two
half-planes, only one of which should be seen by the camera (this is
meant "ideally", if the points and the line are proper fits to the
data). That is to say, each set of inliers should only occupy one
half-plane, and without using the information of which of the two
half-planes is occupied for each plane it is not possible to determine
which angle between the planes one should look at. Therefore
- Romve determineConvexityOfLineFromViewPoint.
- Add function determineConvexityFromViewpointGivenLineAndInlierPoints
that first determines the half-plane occupied by each set of inliers
(via majority-vote) and then determines convexity/concavity based on the
configuration of the two half-planes. Results are identical to those of
determineConvexityFromViewpointGivenLineAndMeanPoints.