Change pipeline for line type assignment + see details
* Change pipeline for line type assignment:
  - First look at convexity/concavity and assign edge type to lines that
have inlier planes that form a convex angle.
  - For the lines whose inlier planes form a concave angle, prolonge the
inlier planes beyond the endpoints of the line. Look at whether each of
these 4 (left/right, beyond each of two endpoints) planes have (enough)
points that fit on it (logical '1') or not (logical '0'). Assign edge
type to the lines that have 4 logical '0's and intersection type to all
the others.
* Reverse "convexity" and "concavity" in the names of the variables in
the code (previously wrong)
* Change parameter for length of prolonged lines