Fix getRectanglesFromLine
For very small lines the edge (among the two parallel to the line)
furthest from the line (e.g., *rect_left[1] -- *rect_left[3]) of the
rectangle could have become closer than the other to the line => Fix
adding offset for this edge as well.

Norm is simply rewritten to use OpenCV function but does not produce any
variations.