Improve visualization (see details) + variables renaming
* Visualization
  - Change order of windows displaying in visualization mode. Now 3D
line with inliers is shown after the 2D line is displayed.
  - Display 2D line also for lines that are later discarded. Necessary:
previously some windows of the 3D line with inliers did not match any
2D-line image.
  - For each line that is successfully projected to 3D display a window
showing the original line/rectangles in 2D, the 2D-reprojection of the
line readjusted through inliers and the prolonged line/rectangles if
any.
* Rename variables for the endpoints of the reprojection of the adjusted
line (previously start, end) to make them more explanatory.