Fix bugs in instance label assignment
- In case of planar/intersection lines, one should prolonge the inlier
planes and count the number of points that are inlier to these prolonged
planes and that have the instance label of either of the two objects on
the sides of the original line. However, to identify the object that if
removed would cause the line to disappear, one should take the least
present of these two instance labels, rather than the most present.
- Fix bug in findInliersWithLabelsGivenPlanes: the first value of the
pair should be taken from cv_cloud_ (that contains x, y, z info) rather
than from instances (that should be always accessed as unsigned short if
labels are in this format).
- Only return *valid* inliers from findInliersWithLabelsGivenPlanes,
rather than all the inliers.
- Remove some debug logs.