Add checks of lines based on 2D info
- Line reprojected in 2D should be at least a certain number of pixels
(default = 2) long to be kept. Solves one case of planar lines not
finding enough inliers in their rectangles because too short.
- Line reprojected in 2D should not be much shorter than the original 2D
line. Solves bug of line with no inliers when labelling in
line_ros_utility::ListenAndPublish::labelLinesWithInstances. There was
in fact at least one case in which the points in the rectangles around
the reprojected 2D line were not inliers to the hessians found in
line_detection::LineDetector::find3DlineOnPlanes. This was because the
latter (hessians) were fitted to the points in the rectangles around the
original 2D line, which contained points not all on the same plane
(meaning that, considering only one side of the line, the points were
not all on the same geometrical plane). In line_ros_utility the
rectangles around the reprojected line contained points only from one of
the two geometrical planes above (from the same side of the line), which
did not coincide with the hessian. Therefore, no inliers were found.