Fix bug of no inliers in rectangles around reprojection + see details
- Fix bug of 3D line finding no inliers in one of the rectangles around
its reprojection in 2D. Apart from the readjustment with inliers, for
which some degenerate case were already handled with fixes, this might
happen also in cases in which no readjustment is done (e.g., for
discontinuity lines). Indeed, when reprojecting a 3D point in 2D with
the image_geometry method project3dToPixel, points are mapped to
strictly integer pixel coordinates, whereas coordinates for the original
2D lines are in general not integer. Therefore, the rectangles might be
slightly different and therefore contain different points. The fix
consists in assuming that all points in the rectangles are inliers when
no inliers are found.
- Remove assignment of null hessian to discontinuity lines. Verified
that this is in fact nor used in the code elsewhere nor justified
anywhere in the code through documentation.
- Move log of lines successfully projected to 3D outside visualization
mode.
- Improve plane drawing in display_line_with_points_and_planes.py
(planes are only drawn in the region where inliers are). matplotlib does
not handle well transparency of the planes, so when plotting both planes
and inliers the visualization is slightly cluttered. Therefore, to plot
planes the corresponding line in the code should be uncommented.