Improve adjustment of discontinuity lines + see details
The previous approach for adjusting discontinuity lines in 3D, based on
finding the nearest point to line, simply did not work, because there
were several cases in which the nearest point was close to the end
endpoint and therefore doing end_guess = nearest_point + direction
yielded a line that did not match the original one. (For an example use
the visualization tools on Frame 5, trajectory 1 of the pyScenenetRGBD
train dataset, to see that the reprojections of the discontinuity lines
were very far away from the correct 2D line).
Current easy solution: take all inlier points in the plane to which the
discontinuity line should be fitted (i.e., the plane closest to the
camera) and take as endpoints of the adjusted 3D line the two inlier
points the projection of which is closer to the endpoints of the 2D
line.