Much better improvement of adjustment of discontinuity lines
Discontinuity lines are now almost perfect fits to the true line, after
adjustment with the following procedure (further details in the comments
throughout the code):
- Take the inlier plane to which the line should be fitted (i.e., the
one closest to the origin) and, using the projection matrix and the
hessian of the plane, find the 3D points on the plane that correspond to
the 2D endpoints of the line.
- Adjust the line by shifting it towards the inliers that in 2D are
closest to the line.