Fix bug of 3D line with endpoint in the origin
A case was found of a 3D line having an endpoint that coincided with the
origin, therefore being mapped to NaN when reprojected in 2D and later
causing SegFault errors. To handle it, simply set the rating of this
lines as very high in findAndRate3DLine, so that it either does not get
selected or it gets discarded because it has too high a rating (cf.
project2Dto3DwithPlanes).