Fix nasty bug due to points in SceneNetRGBD with no depth information
* Some points in the point cloud might have no depth information. In
SceneNetRGBD these are encoded with corresponding {0, 0, 0} coordinates
in the point cloud. If a line is on the edge of a region containing such
points (e.g., frame 248 of trajectory 4 of train dataset 0), it should
be discarded, because planeRANSAC, due to its casuality, might wrongly
select one of these null-coordinate points to fit the plane, therefore
fitting a (completely) wrong plane to a side of the line, causing
several other failures later in the pipeline (e.g. some CHECKs failed or
NaN values were generated). One way to handle this could be simply not
considering the null-coordinate points as inliers, rather than
discarding the entire line, but this way the line could be assigned to a
wrong line type or have remaining inliers that are not descriptive of
the actual plane. We therefore choose to simply discards the line.

Minor changes:
* In fitDiscontLineToInliers:
  - Re-adapt code so as to modify the pointed values only at the end.
  - Add sanity check for NaN points: if this happens simply return the
input line, without readjusting it.
* Small style change in inlier functions.
* Fix typos in comments.