Remove feeding of instance labels in test file
When retrieving the descriptor for a line, the network should not need
the instance label, since in the general case the latter is not
available. This commit removes feeding of labels in the network.

Minor change: adaptation of dataset_name to the format corresponding to
the current version of scenenet_ros_tools