Keep prolonging edge/intersection lines when possible to determine type
It often happens that small line segments are detected, so when looking
at the configuration of inliers in the prolonged planes
(assignEdgeOrIntersectionLineType) both planes, on at least one side,
have inliers. Now the algorithm tries to prolonge the line from this
side/these sides when this happens. Still not perfect.

- Also, [1][1]/[1][1] are now assigned to EDGE, since more edge lines
than intersection lines very found to be in this configuration (that
should never happen and is only due to lines being detected in small
pieces).