Slight change in fitDiscontLineToInliers + other small changes
In fitDiscontLineToInliers: the line is now shifted not towards the
point P that in 2D is closer to the reprojection of the line, but
towards the projection of P onto the inlier plane to which the line must
be fitted. Small changes are expected, as P is itself an inlier to the
inlier plane considered, and is therefore close to the inlier plane (and
therefore to its projection).

Auxiliary changes:
- Remove test for similar length between candidate 3D line and first
guess of 3D line (introduced in 1d21184).
- Undo handling of system() function return values (introduced in
c9e1c50) because it is system specific.
- Add some cv::destroyWindow and handle exceptions.
- Add logs in verbose mode for line_detection.
- Change parameter for max ratio between lines that have similar length
to 3 (line_detection::linesHaveSimilarLength).
- Add checks for non-null pointers in line_detection.h.
- Small style fixes.
- Small improvements in comments (line_detection.h).