Feed geometric info to the neural network
Add two types of parametrization of lines:
* Centerpoint + first two entries of the unit direction vector of the
line: the third entry of the vector can be retrieved enforcing that the
norm of the vector be 1, and is therefore not fed into network. Also, to
have invariance w.r.t. the orientation of the line, the first two
entries of the unit direction vector are adjusted so that the first
entry is strictly non-positive. => 5 parameters per line.
* Orthonormal representation => 4 parameters per line.
The neural network is therefore initialized with a different number of
units in the geometric_info placeholder, accordingly to the type of line
parametrization chosen.

Auxiliary changes:
- Change test files accordingly.
- Add start and end points of the line to the service ImageToEmbeddings.
- Add asserts in python/tools/lines_utils.py.
- Change threshold for check of rotation matrix in
python/tools/lines_utils.py.
- Small style fixes.