Adjust pipeline to allow use of SceneNN
- Adjust generation script (python/generate_trajectory_files.sh) so as
to handle new dataset SceneNN.
- Add SCENENN_DATASET_PATH in python/config_paths_and_variables, to
point to the path where the SceneNN dataset is located.
- Rename detect_cluster_show to detect_and_save_lines.
- Add node scenenn_to_line_tools. Launch either this node or
scenenet_to_line_tools depending on whether the dataset is SceneNetRGBD
or SceneNN.
- Add argument start_frame to line_ros_utility/ListenAndPublish and to
launch file. Assign this index to the first frame received from the ROS
bag. Create util to get the number of frames in a ROS bag.
- Adjust script to create directories (python/create_data_dir.sh) so as
to create folders with indices between start_frame and end_frame (for
SceneNetRGBD these were set to 0 and 299 respectively, but they vary for
SceneNN).
- Adjust python/get_virtual_camera_images.py by creating two different
functions depending on the dataset used. Also adjust
python/split_dataset_with_labels_world.py.
- Create util to read camera-to-world matrices from SceneNN.
- Fix typos. Small style changes.