Properly project depth image to 3D + see details
* Reorganize scenenet_utils.py by creating two auxiliary modules:
cloud_utils.py and virtual_camera_utils.py. Move all functions stricly
related to finding the pose, origin etc. of the virtual camera in
virtual_camera_utils.py and all the functions related to transformation,
reprojection, etc. of the cloud to cloud_utils.py. Also, create a
scenenn_utils.py module parallel to scenenet_utils.py. Each of the
latter now only contains the camera model and the function to reproject
pixels in a depth image to a point cloud. This makes it more clear that
one should distinguish in general between the virtual camera, that we
always take as having the SceneNetRGBD (pinhole) model, and the real
camera (i.e., the model of the one from which the images in the dataset
were taken). The real camera is needed to obtain the point cloud from
the RGB-D image and is therefore different from SceneNetRGBD and
SceneNN.
The same changes are mirrored in line_description.
Also NOTE that SceneNN uses a different depth image compared to
SceneNetRGBD: it directly encodes the z coordinate in the depth
information of each pixel, as opposed to SceneNetRGBD, that uses instead
the ray length.
* Add package pptk to visualize point cloud from which the virtual
camera image is obtained.
* Improve docs and comments in the virtual camera image retrieving part.
* Fix bug in line_description/nodes/virtual_camera_image_retriever.py:
the coloured point cloud had the position and RGB channels inverted.

Minor changes:
- Add check in python/generate_trajectory_files.sh to avoid overwriting
existing tar files.
- Style fixes.