Add check for single connected component of inliers in planeRANSAC
The function that fits planes to the points in a rectangle (planeRANSAC)
often fails to find the correct plane, because the plane that has most
inliers is actually fitted to sets of points that do not form connected
components (e.g. one small 'cluster' of points is closer to the origin
than another small 'cluster' of points that are inliers to the same
plane). This is often due to rectangles being not perfect fits around
the line and is therefore not avoidable, due to intrinsic precision
errors when detecting the line in 2D and fitting the rectangles around
it. However, it is an important issue, in particular to obtain accurate
discontinuity lines, that rely on the planes being good fits to the
points. A possible workaround consists in checking that the inliers to
the plane found form connected components (ideally they should form a
single connected rectangle on a 3D plane, but this is often not the
case). This increases the computational cost of the RANSAC operation,
but is crucial to achieve better accuracy when fitting planes.

Three different structures are implemented to perform the task:
- A union-find structure, that clusters points based on distance, as new
points are detected as inliers.
- A priority-queue-like structure, that instead keeps the sorted
pairwise distances between the points and finds the number of
disconnected components by looking at the discontinuities of the
distribution of the sorted distances (as an example, consider the
histogram of the pairwise sorted distances: when the points are in two
or more different disconnected components the histogram will show two or
more clearly separated modes).
- A octree structure, that organizes points in a 3D-tree structure and
checks whether they all lie in a single connected component by
trasversing the tree from each point to its nearest unvisited neighbours
until all points that are connected (i.e., not further away from one
another than a certain distance) are visited.