Add much more efficient check for single component in planeRANSAC
The new check defines a set of points as being not part of a single
connected component if the distribution of the sorted distances of each
point from the mean point has a discontinuity above a certain threshold.
Let us note that this method as a time complexity linear in the number
of inliers, whereas those in aac314d had a worst-case quadratic
complexity.