Fix for 'batch-really-all' triplet selection
The loss for the easy triplet should still be set to 0, as done
previously than 9c468c1. It is the total loss of the batch that should
be averaged also only the easy loss. Also, the regularization term
changes to also include anchor-positive distances for anchor-positive
pairs that form easy triplets.