Add possibility to use ground-truth instance labels or not in test.py
At test time, the ground-truth instance labels can be used to display
ground-truth labels and to assign each embedding in the checkpoint for
TensorBoard's Projector a label (that can be used, e.g., to colour the
points by instance label). However, in principle, for a generic test
dataset it might be the case that ground-truth instance labels are not
available (e.g., when testing on a real set without ground-truth
labelling). It therefore makes sense not to use them at test time (i.e.,
when just retrieving the embeddings), since they are not required by the
network to obtain the embeddings, but are only used for optimization at
train time.