Add line matching with SIFT-like matching criterion
As in SIFT, a candidate match is considered a valid match only if the
ratio between its rating and the rating of the second-best candidate
match is below a certain threshold (0.8 by default). The entire
line_matching module is not sensible to use when the loss optimizes
distances in the embedding space with the purpose of forming clusters
based on *instance* rather than on *specific line features*. Therefore,
the code from this commit is pushed only to complete the line_matching
module, but is currently not expected to be used. Or, at least, the
matching obtained should not be expected to be line-to-line, but rather
(line of a certain instance)-to-(line of the same instance).