Make triplet loss statistics epoch/batch step a parameter
Also, an important thing was noticed: due to dropout, the loss and
regularization term in the statistics log file will not perfectly match
the values shown in the summary in TensorBoard (unless the dropout rate
is 1.0). This is because, although executed feeding the same batch data,
the train operation and the statistics-retrieval operations need to be
executed separately and therefore the weights in the dropout layer can
happen to be selected differently. Use dropout rate equal to 1.0 to
obtain a perfect match.

Minor changes:
- Move train operation before the statistics operation, so as to have
the statistics up-to-date with the parameters changed by the application
of gradients.