Add statistics about the embeddings retrieved
Display statistics about the embeddings obtained for each instance in a
certain dataset, i.e., it displays statistics about the potential
clusters that can be formed in the embedding space (e.g., mean and
standard deviation of the distances from the 'mean embedding', max and
min 'intra-instance' distance, etc.).

Minor changes:
- Save ground-truth instance labels in the same order as the embeddings
retrieved, that can therefore be assigned to the correct label for the
statistics introduced with the current commit.
- Style fixes.